fix(tests): new tests;
- Moved tests to validators tests.
- Fixed raises and expected errors.
